Add more OOP classes
